+++ tomcat/src/main/java/org/apache/coyote/http11/handler/user/RegisterHandler.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11.handler;
+package org.apache.coyote.http11.handler.user;
 
 import static org.apache.coyote.http11.header.HttpHeaderType.LOCATION;
 import static org.apache.coyote.http11.http.HttpVersion.HTTP11;
@@ -8,6 +8,7 @@ import java.util.Map;
 import nextstep.jwp.application.UserService;
 import nextstep.jwp.dto.UserRegisterRequest;
 import org.apache.catalina.utils.Parser;
+import org.apache.coyote.http11.handler.ResourceHandler;
 import org.apache.coyote.http11.header.HttpHeader;
 import org.apache.coyote.http11.http.request.HttpRequest;
 import org.apache.coyote.http11.http.response.HttpResponse;
