+++ tomcat/src/main/java/org/apache/coyote/http11/handler/user/RegisterPageHandler.java
@@ -1,5 +1,6 @@
-package org.apache.coyote.http11.handler;
+package org.apache.coyote.http11.handler.user;
 
+import org.apache.coyote.http11.handler.ResourceHandler;
 import org.apache.coyote.http11.http.request.HttpRequest;
 import org.apache.coyote.http11.http.response.HttpResponse;
